fix integer bug in getX()

+++ java_example/src/de/thu/h5ptranslate/Element.java
@@ -1,10 +1,10 @@
 package de.thu.h5ptranslate;
 
 public interface Element{
-    int getX();
-    int getY();
-    int getWidth();
-    int getHeight();
+    Float getX();
+    Float getY();
+    Float getWidth();
+    Float getHeight();
     String getText();
     String getContentName();
     String getFile();
